Reorganize files to under organization/groups_admin
Since the entry point already have a bunch of bootstrapping code for
header etc.
