+++ app/views/admin/organizations/groups.html.erb
@@ -7,7 +7,7 @@
     var config = <%= safe_js_object frontend_config %>;
     var user_data = <%= safe_js_object current_user.data.to_json.html_safe %>;
   </script>
-  <%= javascript_include_tag 'cdb.js', 'models.js', 'organization_templates.js', 'organization_deps.js', 'organization_groups.js' -%>
+  <%= javascript_include_tag 'cdb.js', 'models.js', 'organization_templates.js', 'organization_deps.js', 'organization.js' -%>
 <% end %>
 <%= content_for(:css) do %>
   <%= stylesheet_link_tag 'organization.css', :media => 'all' %>
