+++ lib/assets/javascripts/cartodb/common/dialogs/add_group_users/add_group_users_view.js
@@ -4,7 +4,7 @@ var cdb = require('cartodb.js');
 var BaseDialog = require('../../views/base_dialog/view.js');
 var OrgUsersView = require('../../views/org_users/org_users_view');
 var randomQuote = require('../../view_helpers/random_quote.js');
-var GroupUsersView = require('../../../organization_groups/group_users_view');
+var GroupUsersView = require('../../../organization/groups_admin/group_users_view');
 
 /**
  * Dialog to add custom basemap to current map.
