+++ lib/assets/javascripts/cartodb/organization/entry.js
@@ -11,6 +11,8 @@ var ColorPickerView = require('./color_picker_view');
 var OrganizationUsersView = require('./organization_users/organization_users_view');
 var domainRegExp = /^[a-zA-Z0-9][a-zA-Z0-9-_]{0,61}[a-zA-Z0-9]{0,1}\.([a-zA-Z]{1,6}|[a-zA-Z0-9-]{1,30}\.[a-zA-Z]{2,3})$/;
 var RegenerateKeysDialog = require('../keys/regenerate_keys_dialog_view');
+var GroupsRouter = require('./groups_admin/router');
+var GroupsMainView = require('./groups_admin/groups_main_view');
 
 if (window.trackJs) {
   window.trackJs.configure({
@@ -175,7 +177,7 @@ $(function() {
           authenticityToken: window.authenticity_token,
           clean_on_hide: true
         }).appendToBody();
-      })  
+      })
     }
 
     // API keys
@@ -204,6 +206,34 @@ $(function() {
 
       regenerateApiKeyHandler(ev, 'organization_user', '/u/' + current_username + '/organization/users/' + username + '/regenerate_api_key');
     });
+
+    var $groups = $('.js-groups-content');
+    if ($groups) {
+      if (!currentUser.isOrgAdmin()) {
+        window.location = currentUser.viewUrl().accountSettings();
+        return false;
+      }
+
+      var groups = new cdb.admin.OrganizationGroups([], {
+        organization: currentUser.organization
+      });
+      var router = new GroupsRouter({
+        rootUrl: currentUser.viewUrl().organization().groups(),
+        groups: groups,
+        user: currentUser
+      });
+
+      var groupsMainView = new GroupsMainView({
+        el: $groups,
+        groups: groups,
+        router: router,
+        user: currentUser
+      });
+      groupsMainView.render();
+      window.groups = groupsMainView;
+
+      router.enableAfterMainView();
+    }
   });
 
 });
