+++ lib/assets/javascripts/cartodb/organization/groups_admin/create_group.jst.ejs
