+++ lib/assets/javascripts/cartodb/organization/groups_admin/create_group_view.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var cdb = require('cartodb.js');
-var randomQuote = require('../common/view_helpers/random_quote');
+var randomQuote = require('../../common/view_helpers/random_quote');
 
 /**
  * View to create a new group for an organization.
@@ -29,7 +29,7 @@ module.exports = cdb.core.View.extend({
         quote: randomQuote()
       });
     } else {
-      html = this.getTemplate('organization_groups/create_group')({
+      html = this.getTemplate('organization/groups_admin/create_group')({
       });
     }
     this.$el.html(html);
