+++ lib/assets/javascripts/cartodb/organization/groups_admin/edit_group.jst.ejs
