+++ lib/assets/javascripts/cartodb/organization/groups_admin/edit_group_view.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var cdb = require('cartodb.js');
-var randomQuote = require('../common/view_helpers/random_quote');
+var randomQuote = require('../../common/view_helpers/random_quote');
 
 /**
  * View to edit an organization group.
@@ -29,7 +29,7 @@ module.exports = cdb.core.View.extend({
         quote: randomQuote()
       });
     } else {
-      html = this.getTemplate('organization_groups/edit_group')({
+      html = this.getTemplate('organization/groups_admin/edit_group')({
         displayName: this.options.group.get('display_name')
       });
     }
