+++ lib/assets/javascripts/cartodb/organization/groups_admin/group_header.jst.ejs
