+++ lib/assets/javascripts/cartodb/organization/groups_admin/group_header_view.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var cdb = require('cartodb.js');
-var pluralizeString = require('../common/view_helpers/pluralize_string');
+var pluralizeString = require('../../common/view_helpers/pluralize_string');
 
 /**
  * Header view when looking at details of a specific group.
@@ -35,7 +35,7 @@ module.exports = cdb.core.View.extend({
     }
 
     this.$el.html(
-      this.getTemplate('organization_groups/group_header')(d)
+      this.getTemplate('organization/groups_admin/group_header')(d)
     );
     return this;
   },
