+++ lib/assets/javascripts/cartodb/organization/groups_admin/group_user_view.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var cdb = require('cartodb.js');
-var ViewFactory = require('../common/view_factory');
+var ViewFactory = require('../../common/view_factory');
 
 /**
  * View of a single group user.
@@ -22,7 +22,7 @@ module.exports = cdb.core.View.extend({
 
   render: function() {
     this.$el.html(
-      this.getTemplate('organization_groups/group_user_view')({
+      this.getTemplate('organization/groups_admin/group_user_view')({
         avatarUrl: this.model.get('avatar_url'),
         username: this.model.get('username'),
         email: this.model.get('email')
