+++ lib/assets/javascripts/cartodb/organization/groups_admin/group_user_view.jst.ejs
