+++ lib/assets/javascripts/cartodb/organization/groups_admin/group_users_view.js
