+++ lib/assets/javascripts/cartodb/organization/groups_admin/groups_index_view.js
