+++ lib/assets/javascripts/cartodb/organization/groups_admin/groups_main_view.js
@@ -1,7 +1,7 @@
 var $ = require('jquery');
 var _ = require('underscore');
 var cdb = require('cartodb.js');
-var navigateThroughRouter = require('../common/view_helpers/navigate_through_router');
+var navigateThroughRouter = require('../../common/view_helpers/navigate_through_router');
 
 /**
  * Controller view, managing view state of the groups entry point
