+++ lib/assets/javascripts/cartodb/organization/groups_admin/router.js
@@ -1,14 +1,14 @@
 var _ = require('underscore');
 var cdb = require('cartodb.js');
-var Router = require('../common/router');
+var Router = require('../../common/router');
 var RouterModel = require('./router_model');
 var GroupHeaderView = require('./group_header_view');
 var GroupsIndexView = require('./groups_index_view');
 var CreateGroupView = require('./create_group_view');
 var ShowGroupView = require('./show_group_view');
 var EditGroupView = require('./edit_group_view');
-var ViewFactory = require('../common/view_factory');
-var PaginationModel = require('../common/views/pagination/model');
+var ViewFactory = require('../../common/view_factory');
+var PaginationModel = require('../../common/views/pagination/model');
 
 /**
  *  Backbone router for organization groups urls.
