+++ lib/assets/javascripts/cartodb/organization/groups_admin/router_model.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var cdb = require('cartodb.js');
-var ViewFactory = require('../common/view_factory');
-var randomQuote = require('../common/view_helpers/random_quote');
+var ViewFactory = require('../../common/view_factory');
+var randomQuote = require('../../common/view_helpers/random_quote');
 
 /**
  * Model representing the router state
