+++ lib/assets/javascripts/cartodb/organization/groups_admin/show_group_view.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var cdb = require('cartodb.js');
 var GroupUsersView = require('./group_users_view');
-var AddGroupUsersView = require('../common/dialogs/add_group_users/add_group_users_view');
+var AddGroupUsersView = require('../../common/dialogs/add_group_users/add_group_users_view');
 
 /**
  * View representing the default "show" defaults of a group.
