--- lib/assets/javascripts/cartodb/organization_groups.js
@@ -1,36 +1,0 @@
-var $ = require('jquery');
-var cdb = require('cartodb.js');
-var Router = require('./organization_groups/router');
-var GroupsMainView = require('./organization_groups/groups_main_view');
-
-$(function() {
-  cdb.init(function() {
-    cdb.templates.namespace = 'cartodb/';
-    cdb.config.set('url_prefix', window.user_data.base_url);
-    var user = new cdb.admin.User(window.user_data);
-    if (!user.isOrgAdmin()) {
-      window.location = user.viewUrl().accountSettings();
-      return false;
-    }
-
-    var groups = new cdb.admin.OrganizationGroups([], {
-      organization: user.organization
-    });
-    var router = new Router({
-      rootUrl: user.viewUrl().organization().groups(),
-      groups: groups,
-      user: user
-    });
-
-    var groupsMainView = new GroupsMainView({
-      el: $('.js-groups-content'),
-      groups: groups,
-      router: router,
-      user: user
-    });
-    groupsMainView.render();
-    window.groups = groupsMainView;
-
-    router.enableAfterMainView();
-  });
-});
