+++ lib/assets/test/spec/cartodb/organization/groups_admin/create_group_view.spec.js
@@ -1,7 +1,7 @@
 var cdb = require('cartodb.js');
-var CreateGroupView = require('../../../../javascripts/cartodb/organization_groups/create_group_view');
+var CreateGroupView = require('../../../../../javascripts/cartodb/organization/groups_admin/create_group_view');
 
-describe('organization_groups/create_group_view', function() {
+describe('organization/groups_admin/create_group_view', function() {
   beforeEach(function() {
     this.group = new cdb.admin.Group();
     spyOn(this.group, 'save');
