+++ lib/assets/test/spec/cartodb/organization/groups_admin/edit_group_view.spec.js
@@ -1,7 +1,7 @@
 var cdb = require('cartodb.js');
-var EditGroupView = require('../../../../javascripts/cartodb/organization_groups/edit_group_view');
+var EditGroupView = require('../../../../../javascripts/cartodb/organization/groups_admin/edit_group_view');
 
-describe('organization_groups/edit_group_view', function() {
+describe('organization/groups_admin/edit_group_view', function() {
   beforeEach(function() {
     this.group = new cdb.admin.Group({
       id: 'g1',
