+++ lib/assets/test/spec/cartodb/organization/groups_admin/group_header_view.spec.js
@@ -1,7 +1,7 @@
 var cdb = require('cartodb.js');
-var GroupHeaderView = require('../../../../javascripts/cartodb/organization_groups/group_header_view');
+var GroupHeaderView = require('../../../../../javascripts/cartodb/organization/groups_admin/group_header_view');
 
-describe('organization_groups/group_header_view', function() {
+describe('organization/groups_admin/group_header_view', function() {
 
   beforeEach(function() {
     this.setupView = function() {
