+++ lib/assets/test/spec/cartodb/organization/groups_admin/groups_main_view.spec.js
@@ -1,8 +1,8 @@
 var $ = require('jquery');
 var cdb = require('cartodb.js');
-var GroupsMainView = require('../../../../javascripts/cartodb/organization_groups/groups_main_view');
+var GroupsMainView = require('../../../../../javascripts/cartodb/organization/groups_admin/groups_main_view');
 
-describe('organization_groups/groups_main_view', function() {
+describe('organization/groups_admin/groups_main_view', function() {
 
   beforeEach(function() {
     this.contentView = new cdb.core.View();
