+++ lib/assets/test/spec/cartodb/organization/groups_admin/router.spec.js
@@ -1,7 +1,7 @@
 var cdb = require('cartodb.js');
-var Router = require('../../../../javascripts/cartodb/organization_groups/router');
+var Router = require('../../../../../javascripts/cartodb/organization/groups_admin/router');
 
-describe("organization_groups/router", function() {
+describe("organization/groups_admin/router", function() {
   beforeEach(function() {
     var user = new cdb.admin.User({
       id: 123,
