+++ lib/assets/test/spec/cartodb/organization/groups_admin/show_group_view.spec.js
@@ -1,9 +1,9 @@
 var $ = require('jquery');
 var cdb = require('cartodb.js');
-var ShowGroupView = require('../../../../javascripts/cartodb/organization_groups/show_group_view');
-var AddGroupUsersView = require('../../../../javascripts/cartodb/common/dialogs/add_group_users/add_group_users_view.js');
+var ShowGroupView = require('../../../../../javascripts/cartodb/organization/groups_admin/show_group_view');
+var AddGroupUsersView = require('../../../../../javascripts/cartodb/common/dialogs/add_group_users/add_group_users_view.js');
 
-describe('organization_groups/show_group_view', function() {
+describe('organization/groups_admin/show_group_view', function() {
 
   beforeEach(function() {
     this.user = new cdb.admin.User({
